requirements.txt
telethon
cryptg
heroku3
psycopg2
requests
Pillow
telegraph
